[Frontend] Consolidate entrypoint middleware (#52309)
