+++ tests/entrypoints/serve/middleware/test_optional_middleware.py
