+++ vllm/entrypoints/speech_to_text/factories.py
@@ -37,12 +37,6 @@ def register_speech_to_text_api_routers(
         app.include_router(translation_router)
 
 
-def add_websocket_metrics_middleware(app: FastAPI):
-    from .realtime.metrics import WebSocketMetricsMiddleware
-
-    app.add_middleware(WebSocketMetricsMiddleware)
-
-
 def init_speech_to_text_state(
     engine_client: "EngineClient",
     state: "State",
